requirements.txt
openai
click
python-dotenv
jinja2
requests
pyyaml
fastapi
uvicorn
